add checksum ...checking
