+++ .circleci/config.yml
@@ -261,6 +261,11 @@ jobs:
       - attach_workspace:
           at: .
       - run:
+          name: Audit checksums of color files
+          command: |
+            chmod 744 ./checkcolorhash.sh
+            ./checkcolorhash.sh;
+      - run:
           name: Audit vulnerability of frontend node_modules
           command: |
             cd frontend
