+++ frontend/src/colors.scss
@@ -1,9 +1,8 @@
 
 // STOP! Don't edit this file.
 // Instead, edit colors.js and run 'yarn makecolors'
-// or if you are using docker, yarn docker:makecolors from the root directory
-// from the frontend directory to generate this file'
-// (the colors variables are used in two ways, inlined in javascript and directly in css
+// or if you are using docker, yarn docker:makecolors
+// to generate this file (the colors variables are used in two ways, inlined in javascript and directly in css
 // to keep things consistent, we update only in one place - the colors.js file, and use
 // yarn makecolors to keep the files identical)
